+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # redb - Changelog
 
+## Unreleased
+* Optimize `Table::pop_first()` and `Table::pop_last()` to be about 2x faster.
+
 ## 4.1.0 - 2026-04-19
 **This release contains a large number of bug fixes discovered by AI coding agents**